Moved some code up to WebSocketClientThreadBase
